requirements.txt
py-cord
deepgram-sdk
pynacl
python-dotenv
openai